working on target/lick/behavior models

+++ nems/signal.py
@@ -340,7 +340,7 @@ class SignalBase:
                 raise ValueError(m)
             mask = self.epochs['name'] == epoch
             bounds = self.epochs.loc[mask, ['start', 'end']].values
-            bounds = np.round(bounds * self.fs) / self.fs
+            bounds = np.round(bounds.astype(float) * self.fs) / self.fs
 
         if boundary_mode is None:
             raise NotImplementedError
